feat(adapters): add Pandas and JSON file adapters for data conversion
